time setting auto reset issur solved for expiration time selector
