+++ src/front_end/fx_maket_watch/lib/screens/user_screen/alert_form.dart
@@ -8,6 +8,7 @@ import '../../blocs/edit_alert_bloc/edit_alert_bloc.dart';
 import 'alert_condition.dart';
 import 'currency_nd_timeframe.dart';
 import '../../services/api_calls.dart';
+import 'expiration_selector.dart';
 import 'price_and_alert_medium.dart';
 
 class AlertForm extends StatelessWidget {
@@ -23,79 +24,19 @@ class AlertForm extends StatelessWidget {
           AlertCondition(),
           PriceAndMedium(),
           ExpirationSelector(),
+          RepeatSetpointAndNote(),
         ],
       ),
     );
   }
 }
 
-class ExpirationSelector extends StatelessWidget {
-  const ExpirationSelector({super.key});
+class RepeatSetpointAndNote extends StatelessWidget {
+  const RepeatSetpointAndNote({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Card(
-      child: Container(
-        width: 600,
-        padding: EdgeInsets.symmetric(vertical: 5.0, horizontal: 20.0),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Column(
-              children: [
-                const Text("Expiration Date"),
-                BlocBuilder<EditAlertBloc, EditAlertState>(
-                  builder: (context, state) {
-                    final mycontext = context.read<EditAlertBloc>();
-                    return OutlinedButton(
-                      onPressed: () async {
-                        var selectedDate = await showDatePicker(
-                          context: context,
-                          initialDate: DateTime.now(),
-                          firstDate: DateTime.now(),
-                          lastDate:
-                              DateTime.now().add(const Duration(days: 365)),
-                        );
-                        if (selectedDate != null) {
-                          mycontext.add(
-                              UpdateExpiration(selectedDate: selectedDate));
-                        }
-                      },
-                      child: Text(state.myAlert.expirationInDate != null
-                          ? "${state.myAlert.expirationInDate?.month.toString()}/${state.myAlert.expirationInDate?.day.toString()}/${state.myAlert.expirationInDate?.year.toString()}"
-                          : "MM/dd/YY"),
-                    );
-                  },
-                )
-              ],
-            ),
-            Column(
-              children: [
-                const Text("Expiration Time"),
-                BlocBuilder<EditAlertBloc, EditAlertState>(
-                  builder: (context, state) {
-                    return OutlinedButton(
-                      onPressed: () async {
-                        var selectedTime = await showTimePicker(
-                          context: context,
-                          initialTime: TimeOfDay.now(),
-                        );
-
-                        if (selectedTime != null) {
-                          print(selectedTime);
-                        } else {
-                          print("wahala");
-                        }
-                      },
-                      child: const Text("7:45AM"),
-                    );
-                  },
-                )
-              ],
-            ),
-          ],
-        ),
-      ),
-    );
+    return 
   }
 }
+
